refactor: Post/index.tsx로 수정
